+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 ### Added
 
 - Use validation logic from the app library.
+- Include restrictions data from app metadata files in appcatalogentry CRs.
 
 ## [2.7.0] - 2020-11-09